# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API to create movies",
        "contact": {
            "email": "icaro.barreto166@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/the-movie-api": {
            "get": {
                "tags": [
                    "TMDB"
                ],
                "summary": "Get movie list from tmdb api",
                "description": "Returns list of Movie from tmdb api",
                "operationId": "indexTMDB",
                "parameters": [
                    {
                        "parameter": "Movie--title",
                        "name": "title",
                        "in": "query",
                        "description": "title of Movie",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "titanic"
                        }
                    },
                    {
                        "parameter": "Movie--page",
                        "name": "page",
                        "in": "query",
                        "description": "page of TMDB api",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "title": "success",
                                            "description": "success find movies",
                                            "type": "boolean",
                                            "readOnly": "true",
                                            "example": "true"
                                        },
                                        "data": {
                                            "title": "data",
                                            "type": "array",
                                            "items": {
                                                "example": {
                                                    "page": 1,
                                                    "result": [
                                                        {
                                                            "id": 1,
                                                            "title": "Titanic",
                                                            "release_date": "2002-07-04",
                                                            "overview": "Um artista pobre e uma jovem rica se conhecem e se apaixonam na fatídica jornada do Titanic.",
                                                            "genre": [
                                                                1,
                                                                2
                                                            ]
                                                        }
                                                    ],
                                                    "total_pages": 1
                                                }
                                            }
                                        },
                                        "message": {
                                            "title": "message",
                                            "description": "success message",
                                            "type": "string",
                                            "readOnly": "true",
                                            "example": "sucesso."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Movie not found"
                    }
                }
            }
        }
    },
...
}